refactor: mickey
