+++ app/mickey/constants.ts
@@ -1,0 +1,3 @@
+export const USERS_BY_TEAM = 2;
+export const VISITS_BY_DAY = 4;
+export const WEEK_DAYS = 4;
